feat: lazy-load smart-order-router
