+++ src/test-utils/constants.ts
@@ -1,4 +1,6 @@
 import { ChainId, CurrencyAmount, Percent, Token, TradeType } from '@uniswap/sdk-core'
+// This is a test file, so the import of smart-order-router is allowed.
+// eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import { V3Route } from '@uniswap/smart-order-router'
 import { FeeAmount, Pool } from '@uniswap/v3-sdk'
 import { ZERO_PERCENT } from 'constants/misc'
